+++ Tools/Internal/UnityObjectEditor.cs
@@ -1,0 +1,34 @@
+﻿#if UNITY_EDITOR
+namespace MyBox.Internal
+{
+    using UnityEditor;
+    using UnityEngine;
+    
+    [CustomEditor(typeof(Object), true), CanEditMultipleObjects]
+    public class UnityObjectEditor : Editor
+    {
+        private FoldoutAttributeHandler _foldout;
+        private ButtonMethodHandler _buttonMethod; 
+        
+        private void OnEnable()
+        {
+            _foldout = new FoldoutAttributeHandler(target, serializedObject);
+            _buttonMethod = new ButtonMethodHandler(target);
+        }
+
+        private void OnDisable()
+        {
+            _foldout.OnDisable();
+        }
+
+        public override void OnInspectorGUI()
+        {
+            _foldout.Update();
+            if (!_foldout.OverrideInspector) base.OnInspectorGUI();
+            else _foldout.OnInspectorGUI();
+
+            _buttonMethod.OnInspectorGUI();
+        }
+    }
+}
+#endif