+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+- Added: FoldoutAttribute. Thanks to PixeyeHQ!
 - Fix: TransformShakeExtension critical bug fixed
 
 ## [1.2.0] - 2019-11-13